+++ Lec-10 Admin Pannel With PassortJS/controllers/subCategoryController.js
@@ -1,5 +1,6 @@
 const category = require("../models/CategoryModel");
 const subCategory = require("../models/SubCategoryModel");
+const extraCategory = require("../models/ExtraCategoryModel");
 
 // Add SubCategory Page
 const addSubCategoryPage = async (req, res) => {
@@ -68,16 +69,24 @@ const deleteSubCategory = async (req, res) => {
   console.log("Delete SubCategory Id", deleteId);
 
   try {
-    const deleteSubCategory = await subCategory.findByIdAndDelete(deleteId);
-    console.log(deleteSubCategory);
+    const deletExtraCategory = await extraCategory.deleteMany({
+      subCategory_id: deleteId,
+    });
 
-    if (deleteSubCategory) {
-      req.flash(
-        "success",
-        `${deleteSubCategory.subcategory_title} deleted successfully...`
-      );
+    if (deletExtraCategory) {
+      const deleteSubCategory = await subCategory.findByIdAndDelete(deleteId);
+      console.log(deleteSubCategory);
+
+      if (deleteSubCategory) {
+        req.flash(
+          "success",
+          `${deleteSubCategory.subcategory_title} deleted successfully...`
+        );
+      } else {
+        req.flash("error", "SubCategory Deletion failed...");
+      }
     } else {
-      req.flash("error", "SubCategory Deletion failed...");
+      req.flash("error", "Deletion Failed....");
     }
 
     res.redirect("back");
@@ -87,9 +96,55 @@ const deleteSubCategory = async (req, res) => {
   }
 };
 
+// Update SubCategory Page
+const updateSubCategoryPage = async (req, res) => {
+  try {
+    const allCategory = await category.find({});
+    const updateSubCategory = await subCategory.findById(req.params.id);
+
+    allCategory && updateSubCategory
+      ? res.render("subcategory/editSubCategory", {
+          allCategory,
+          updateSubCategory,
+          success: "",
+          error: "",
+        })
+      : res.redirect("back");
+  } catch (e) {
+    console.log(e);
+    res.redirect("back");
+  }
+};
+
+// Update SubCategory
+const updateSubCategory = async (req, res) => {
+  console.log(req.body);
+  console.log(req.params.id);
+
+  try {
+    const updateData = await subCategory.findByIdAndUpdate(
+      req.params.id,
+      req.body
+    );
+
+    if (updateData) {
+      req.flash("success", "SubCateory is updated...");
+    } else {
+      req.flash("error", "SubCateory is not updated...");
+    }
+    res.redirect("/subcategory/viewSubCategoryPage");
+  } catch (e) {
+    console.log(e);
+
+    res.redirect("back");
+  }
+};
+
 module.exports = {
   addSubCategoryPage,
   insertSubCategory,
   viewSubCategoryPage,
   deleteSubCategory,
+  updateSubCategoryPage,
+  updateSubCategory,
 };
